+++ code/utils/flaskUtils.py
@@ -1,0 +1,65 @@
+from flask import Flask, request, render_template
+from LEDUtils import *
+
+#Setup the flask object and get it going
+app = Flask(__name__)
+
+# Flask Index
+@app.route("/", methods=['GET'])
+def index():
+	return render_template('index.html')
+
+# Flask BG Color API
+@app.route("/api/bgcolor", methods=['POST'])
+def setBGColor():
+	# Read the values from the POST
+	program = request.form['color']
+	red = int(request.form['red'])
+	green = int(request.form['green'])
+	blue = int(request.form['blue'])
+	
+	# Change the bg color accordingly
+	if program == "solid":
+		LED.bgColor = ["solid", Color(red,green,blue)]
+	else:
+		LED.bgColor = ["animation", program]
+	return ""
+
+# Flask Text Color API
+@app.route("/api/textcolor", methods=['POST'])
+def setTextColor():
+	# Read the values from the POST
+	program = request.form['color']
+	red = int(request.form['red'])
+	green = int(request.form['green'])
+	blue = int(request.form['blue'])
+	
+	# Change the bg color accordingly
+	if program == "solid":
+		LED.textColor = ["solid", Color(red,green,blue)]
+		LED.changeTextColor(LED.textColor[1])
+	else:
+		LED.textColor = ["animation", program]
+	return ""
+
+	# Flask Font API
+@app.route("/api/font", methods=['POST'])
+def setFont():
+	global hoursPrev
+	global minsPrev
+	global secsPrev
+
+	# Read the values from the POST
+	font = request.form['font']
+	
+	# Assign the font variable in LED class
+	LED.font = font
+	print "changing font to", LED.font
+
+	# Reset the background and variables to make the screen refresh completely on next loop iteration
+	LED.colorFill(LED.bgColor[1], True)		# True boolean to make sure that the whole screen wipes including text
+	secsPrev = 99
+	minsPrev = 99
+	hoursPrev = 99
+	return ""
+
